backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
pydantic==2.10.4
python-multipart==0.0.20
openai==1.59.5
tavily-python==0.5.0
python-dotenv==1.0.1
httpx==0.28.1
greenlet>=3.0.0
